Use GenuineCheck in GenuineCheckModal
